+++ tests/random_ladder_test.py
@@ -1,0 +1,5 @@
+"""Test for RandomLadder class"""
+
+from ladder.random_ladder import RandomLadder
+
+rlad = RandomLadder()
